requirements.txt
torch==2.2.1
pytorch-forecasting==1.0.0
pytorch-lightning==2.2.0
lightgbm==4.3.0
prophet==1.1.5
mlflow==2.11.3
optuna==3.6.0
evidently==0.4.17
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
pydantic==2.6.3
pyyaml==6.0.1
click==8.1.7
holidays==0.44
statsmodels==0.14.1
plotly==5.19.0
meteostat==1.6.7
requests==2.31.0
joblib==1.3.2
scipy==1.12.0

# Development dependencies
ruff==0.3.3
black==24.2.0
isort==5.13.2
pytest==8.1.1
pytest-cov==4.1.0
pre-commit==3.6.2
jupyter==1.0.0
matplotlib==3.8.3
seaborn==0.13.2
